Use size_t in work methods

+++ gr-blocks/lib/pdu_to_tagged_stream_impl.cc
@@ -69,7 +69,7 @@ namespace gr {
     }
 
     int
-    pdu_to_tagged_stream_impl::work (int noutput_items,
+    pdu_to_tagged_stream_impl::work (size_t noutput_items,
                                      gr_vector_int &ninput_items,
                                      gr_vector_const_void_star &input_items,
                                      gr_vector_void_star &output_items)
