+++ gr-blocks/lib/pdu_to_tagged_stream_impl.h
@@ -40,7 +40,7 @@ namespace gr {
 
       int calculate_output_stream_length(const gr_vector_int &ninput_items);
 
-      int work(int noutput_items,
+      int work(size_t noutput_items,
           gr_vector_int &ninput_items,
           gr_vector_const_void_star &input_items,
           gr_vector_void_star &output_items
