+++ gr-digital/lib/packet_headergenerator_bb_impl.cc
@@ -80,7 +80,7 @@ namespace gr {
     }
 
     int
-    packet_headergenerator_bb_impl::work (int noutput_items,
+    packet_headergenerator_bb_impl::work (size_t noutput_items,
                        gr_vector_int &ninput_items,
                        gr_vector_const_void_star &input_items,
                        gr_vector_void_star &output_items)
