+++ gr-fec/lib/tagged_encoder_impl.h
@@ -42,7 +42,7 @@ namespace gr {
                           int mtu=1500);
       ~tagged_encoder_impl();
 
-      int work(int noutput_items,
+      int work(size_t noutput_items,
                gr_vector_int& ninput_items,
                gr_vector_const_void_star &input_items,
                gr_vector_void_star &output_items);
